requirements.txt
django
psycopg2
pillow
django-mptt
django-allauth
django-phone-field